+++ source/rtcw_mp/src/botlib/be_aas_route.c
@@ -1975,7 +1975,7 @@ void AAS_DecompressVis( byte *in, int numareas, byte *decompressed ) {
 
 	//row = (numareas+7)>>3;
 	out = decompressed;
-	end = ( byte * )( (int)decompressed + numareas );
+	end = ( byte * )( (intptr_t)decompressed + numareas );
 
 	do
 	{
