+++ source/rtcw_sp/src/botlib/be_ai_goal.c
@@ -137,7 +137,7 @@ typedef struct iteminfo_s
 	int number;                         //number of the item info
 } iteminfo_t;
 
-#define ITEMINFO_OFS( x ) (int)&( ( (iteminfo_t *)0 )->x )
+#define ITEMINFO_OFS( x ) (intptr_t)&( ( (iteminfo_t *)0 )->x )
 
 fielddef_t iteminfo_fields[] =
 {
