+++ source/rtcw_sp/src/server/sv_bot.c
@@ -546,7 +546,7 @@ BotImport_AICast_VisibleFromPos
 */
 qboolean BotImport_AICast_VisibleFromPos(   vec3_t srcpos, int srcnum,
 											vec3_t destpos, int destnum, qboolean updateVisPos ) {
-	return VM_Call( gvm, AICAST_VISIBLEFROMPOS, (int)srcpos, srcnum, (int)destpos, destnum, updateVisPos );
+	return VM_Call( gvm, AICAST_VISIBLEFROMPOS, (intptr_t)srcpos, srcnum, (intptr_t)destpos, destnum, updateVisPos );
 }
 
 /*
@@ -555,7 +555,7 @@ BotImport_AICast_CheckAttackAtPos
 ===============
 */
 qboolean BotImport_AICast_CheckAttackAtPos( int entnum, int enemy, vec3_t pos, qboolean ducking, qboolean allowHitWorld ) {
-	return VM_Call( gvm, AICAST_CHECKATTACKATPOS, entnum, enemy, (int)pos, ducking, allowHitWorld );
+	return VM_Call( gvm, AICAST_CHECKATTACKATPOS, entnum, enemy, (intptr_t)pos, ducking, allowHitWorld );
 }
 // done.
